refactor: rename apps/repository to apps/repositories for consistency
